split setting.go as multiple files
