+++ modules/setting/webhook.go
@@ -1,0 +1,30 @@
+// Copyright 2019 The Gitea Authors. All rights reserved.
+// Use of this source code is governed by a MIT-style
+// license that can be found in the LICENSE file.
+
+package setting
+
+var (
+	// Webhook settings
+	Webhook = struct {
+		QueueLength    int
+		DeliverTimeout int
+		SkipTLSVerify  bool
+		Types          []string
+		PagingNum      int
+	}{
+		QueueLength:    1000,
+		DeliverTimeout: 5,
+		SkipTLSVerify:  false,
+		PagingNum:      10,
+	}
+)
+
+func newWebhookService() {
+	sec := Cfg.Section("webhook")
+	Webhook.QueueLength = sec.Key("QUEUE_LENGTH").MustInt(1000)
+	Webhook.DeliverTimeout = sec.Key("DELIVER_TIMEOUT").MustInt(5)
+	Webhook.SkipTLSVerify = sec.Key("SKIP_TLS_VERIFY").MustBool()
+	Webhook.Types = []string{"gitea", "gogs", "slack", "discord", "dingtalk"}
+	Webhook.PagingNum = sec.Key("PAGING_NUM").MustInt(10)
+}
